@@ -17,18 +17,19 @@ const (
 	TypeMongo    = "mongo"
 )
 
+// DatastoreQuery holds a prepared query statement and their parameters
 type DatastoreQuery struct {
-	Statement  interface{}
-	Parameters []interface{}
+	Statement  any
+	Parameters []any
 }
 
 // Datastore is the interface that maps a generic designed AST returned by translate.AstTranslator to a native query-statement using the passed data.DatastoreTranslator and execute by the passed data.DatastoreExecutor.
 // This should be generally done by translating the Query-AST into the datastore's native query language. If the query returns any result, the datastore should return true, otherwise false.
 type Datastore interface {
-	// Please note that Configure has to be called once before the component can be used (Otherwise Execute() will return an error)!
+	// Configure has to be called once before the component can be used (Otherwise Execute will return an error)!
 	Configure(appConf *configs.AppConfig, alias string) error
 
-	// Execute() translates the given Query-AST into a datastore's native query and executes the query afterwards via the passed data.DatastoreExecutor.
+	// Execute translates the given Query-AST into a datastore's native query and executes the query afterward via the passed data.DatastoreExecutor.
 	Execute(ctx context.Context, query Node) (bool, error)
 }
 
@@ -36,10 +37,10 @@ type Datastore interface {
 // This should be generally done by translating the Query-AST into the datastore's native query language.
 type DatastoreTranslator interface {
 
-	// Please note that Configure has to be called once before the component can be used (Otherwise Execute() will return an error)!
+	// Configure has to be called once before the component can be used (Otherwise Execute will return an error)!
 	Configure(appConf *configs.AppConfig, alias string) error
 
-	// Execute() translates the given Query-AST into a datastore's native query
+	// Execute translates the given Query-AST into a datastore's native query
 	Execute(ctx context.Context, query Node) (DatastoreQuery, error)
 }
 
@@ -49,13 +50,13 @@ type DatastoreTranslator interface {
 // Please note, that connection-pool-handling should also be done by the DatastoreExecutor internally.
 type DatastoreExecutor interface {
 
-	// Configure() waits for the attached databases to be reachable by pinging the configured connection every 3 seconds for 1 minute
+	// Configure waits for the attached databases to be reachable by pinging the configured connection every 3 seconds for 1 minute
 	// and then configures the Datastore itself.
 	//
 	// If the database isn't reachable or the Datastore itself fails during this process, the encountered error will be returned (otherwise nil).
 	Configure(appConf *configs.AppConfig, alias string) error
 
-	// Execute() executes the native query and returns true if the query is not empty and false otherwise.
+	// Execute executes the native query and returns true if the query is not empty and false otherwise.
 	Execute(ctx context.Context, query DatastoreQuery) (bool, error)
 }
 
@@ -67,9 +68,9 @@ type DatastoreExecutor interface {
 // Please note, that the mapped function should include the call-operand as well!
 type CallOpMapper interface {
 
-	// Get the call-operand this handler handles (i.e. 'abs').
+	// Handles gets the call-operand this handler handles (i.e. 'abs').
 	Handles() string
 
-	// Return the corresponding datastore native function (i.e. 'ABS(<arguments>)').
+	// Map returns the corresponding datastore native function (i.e. 'ABS(<arguments>)').
 	Map(args ...string) (string, error)
 }
