@@ -8,8 +8,10 @@ import (
 	"google.golang.org/grpc"
 )
 
-type SpanFunction func(ctx context.Context, args ...interface{}) (interface{}, error)
+// SpanFunction represents a function, which will be wrapped in a tracing span
+type SpanFunction func(ctx context.Context, args ...any) (any, error)
 
+// MetricsProvider is able to record and publish metrics
 type MetricsProvider interface {
 	// Configure metrics provider
 	Configure(ctx context.Context) error
@@ -20,15 +22,16 @@ type MetricsProvider interface {
 	// GetGrpcServerInterceptor Get an interceptor to instrument grpc calls
 	GetGrpcServerInterceptor() grpc.UnaryServerInterceptor
 	// UpdateHistogramMetric write a value with labels to a histogram
-	UpdateHistogramMetric(ctx context.Context, metric constants.MetricInstrument, value interface{}, labels map[string]string)
+	UpdateHistogramMetric(ctx context.Context, metric constants.MetricInstrument, value any, labels map[string]string)
 	// UpdateGaugeMetric write a value with labels to a gauge
-	UpdateGaugeMetric(ctx context.Context, metric constants.MetricInstrument, value interface{}, labels map[string]string)
+	UpdateGaugeMetric(ctx context.Context, metric constants.MetricInstrument, value any, labels map[string]string)
 	// UpdateCounterMetric Increment a counter by a value
-	UpdateCounterMetric(ctx context.Context, metric constants.MetricInstrument, value interface{}, labels map[string]string)
+	UpdateCounterMetric(ctx context.Context, metric constants.MetricInstrument, value any, labels map[string]string)
 	// Shutdown Gracefully shutdown
 	Shutdown(ctx context.Context)
 }
 
+// TraceProvider is able to record and publish traces
 type TraceProvider interface {
 	// Configure trace provider
 	Configure(ctx context.Context) error
@@ -37,9 +40,9 @@ type TraceProvider interface {
 	// GetGrpcServerInterceptor Get an interceptor to instrument grpc calls
 	GetGrpcServerInterceptor() grpc.UnaryServerInterceptor
 	// ExecuteWithRootSpan starts a new root span and executes the function
-	ExecuteWithRootSpan(ctx context.Context, function SpanFunction, spanName string, labels map[string]string, args ...interface{}) (interface{}, error)
+	ExecuteWithRootSpan(ctx context.Context, function SpanFunction, spanName string, labels map[string]string, args ...any) (any, error)
 	// ExecuteWithChildSpan starts a new child span and executes the function
-	ExecuteWithChildSpan(ctx context.Context, function SpanFunction, spanName string, labels map[string]string, args ...interface{}) (interface{}, error)
+	ExecuteWithChildSpan(ctx context.Context, function SpanFunction, spanName string, labels map[string]string, args ...any) (any, error)
 	// Shutdown Gracefully shutdown
 	Shutdown(ctx context.Context)
 }
