@@ -9,37 +9,43 @@ import (
 	"github.com/open-policy-agent/opa/topdown/builtins"
 	"github.com/open-policy-agent/opa/types"
 	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 	"github.com/unbasical/kelon/pkg/constants/logging"
 )
 
 // nolint:gochecknoglobals,gocritic
 var (
+	// logInfo is a buildin log function for rego, logging on the INFO level
 	logInfo = &rego.Function{
 		Name: "log_info",
 		Decl: types.NewVariadicFunction(nil, types.A, nil),
 	}
+	// logDebug is a buildin log function for rego, logging on the DEBUG level
 	logDebug = &rego.Function{
 		Name: "log_debug",
 		Decl: types.NewVariadicFunction(nil, types.A, nil),
 	}
+	// logWarn is a buildin log function for rego, logging on the WARN level
 	logWarn = &rego.Function{
 		Name: "log_warn",
 		Decl: types.NewVariadicFunction(nil, types.A, nil),
 	}
+	// logError is a buildin log function for rego, logging on the ERROR level
 	logError = &rego.Function{
 		Name: "log_error",
 		Decl: types.NewVariadicFunction(nil, types.A, nil),
 	}
+	// logFatal is a buildin log function for rego, logging on the FATAL level
 	logFatal = &rego.Function{
 		Name: "log_fatal",
 		Decl: types.NewVariadicFunction(nil, types.A, nil),
 	}
 )
 
-func makeBuiltinLogFuncForLevel(level log.Level) func(bctx rego.BuiltinContext, terms []*ast.Term) (*ast.Term, error) {
-	return func(bctx rego.BuiltinContext, terms []*ast.Term) (*ast.Term, error) {
-		if !log.IsLevelEnabled(level) {
+// makeBuiltinLogFuncForLevel constructs a OPA buildin function, which logs the provided operands for the provided logrus.Level
+func makeBuiltinLogFuncForLevel(level logrus.Level) func(bctx rego.BuiltinContext, terms []*ast.Term) (*ast.Term, error) {
+	return func(_ rego.BuiltinContext, terms []*ast.Term) (*ast.Term, error) {
+		if !logrus.IsLevelEnabled(level) {
 			return ast.NullTerm(), nil
 		}
 
@@ -57,7 +63,8 @@ func makeBuiltinLogFuncForLevel(level log.Level) func(bctx rego.BuiltinContext, 
 	}
 }
 
-func logWithLevel(logLevel log.Level, operands *ast.Array) error {
+// logWithLevel tries to log all operands as a single line with the given logrus.Level
+func logWithLevel(logLevel logrus.Level, operands *ast.Array) error {
 	buf := make([]string, operands.Len())
 	idx := 0
 
@@ -78,15 +85,15 @@ func logWithLevel(logLevel log.Level, operands *ast.Array) error {
 	}
 
 	switch logLevel {
-	case log.InfoLevel:
+	case logrus.InfoLevel:
 		logging.LogForComponent("policy").Info(strings.Join(buf, " "))
-	case log.DebugLevel:
+	case logrus.DebugLevel:
 		logging.LogForComponent("policy").Debug(strings.Join(buf, " "))
-	case log.WarnLevel:
+	case logrus.WarnLevel:
 		logging.LogForComponent("policy").Warn(strings.Join(buf, " "))
-	case log.ErrorLevel:
+	case logrus.ErrorLevel:
 		logging.LogForComponent("policy").Error(strings.Join(buf, " "))
-	case log.FatalLevel:
+	case logrus.FatalLevel:
 		logging.LogForComponent("policy").Error(strings.Join(buf, " "))
 		return topdown.Halt{Err: errors.Errorf("Fatal Log: %s", strings.Join(buf, " "))}
 	default:
