@@ -4,40 +4,8 @@ import (
 	"net/http"
 )
 
-type InMemResponseWriter struct {
-	body       []byte
-	statusCode int
-	header     http.Header
-}
-
-func NewInMemResponseWriter() *InMemResponseWriter {
-	return &InMemResponseWriter{
-		header: http.Header{},
-	}
-}
-
-func (w *InMemResponseWriter) Header() http.Header {
-	return w.header
-}
-
-func (w *InMemResponseWriter) Body() string {
-	return string(w.body)
-}
-
-func (w *InMemResponseWriter) StatusCode() int {
-	return w.statusCode
-}
-
-func (w *InMemResponseWriter) Write(b []byte) (int, error) {
-	w.body = b
-	// implement it as per your requirement
-	return 0, nil
-}
-
-func (w *InMemResponseWriter) WriteHeader(statusCode int) {
-	w.statusCode = statusCode
-}
-
+// PassThroughResponseWriter wraps a http.ResponseWriter but the status code and body can be accessed.
+// This is useful for e.g. logging and tracing
 type PassThroughResponseWriter struct {
 	body       []byte
 	statusCode int
@@ -45,6 +13,7 @@ type PassThroughResponseWriter struct {
 	writer     http.ResponseWriter
 }
 
+// NewPassThroughResponseWriter wraps the given http.ResponseWriter
 func NewPassThroughResponseWriter(w http.ResponseWriter) *PassThroughResponseWriter {
 	return &PassThroughResponseWriter{
 		header: http.Header{},
@@ -52,23 +21,26 @@ func NewPassThroughResponseWriter(w http.ResponseWriter) *PassThroughResponseWri
 	}
 }
 
+// Header returns the response http.Header
 func (w *PassThroughResponseWriter) Header() http.Header {
 	return w.writer.Header()
 }
 
-func (w *PassThroughResponseWriter) Body() string {
-	return string(w.body)
-}
+// Body returns the response body
+func (w *PassThroughResponseWriter) Body() []byte { return w.body }
 
+// StatusCode returns the response status code
 func (w *PassThroughResponseWriter) StatusCode() int {
 	return w.statusCode
 }
 
+// Write records the data and passes the buffer to the underlying http.ResponseWriter
 func (w *PassThroughResponseWriter) Write(b []byte) (int, error) {
-	w.body = b
+	w.body = append(w.body, b...)
 	return w.writer.Write(b)
 }
 
+// WriteHeader records the statusCode and passes it on to the underlying http.ResponseWriter
 func (w *PassThroughResponseWriter) WriteHeader(statusCode int) {
 	w.statusCode = statusCode
 	w.writer.WriteHeader(statusCode)
